% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_diag.R
\name{corfun}
\alias{corfun}
\title{Function for evaluating model fit.}
\usage{
corfun(y, yhat, type = c("gaussian", "binomial"))
}
\arguments{
\item{y}{Dependent variable used in the model.}

\item{yhat}{Predictions of the dependent variable.}

\item{type}{Character string giving the type of model used (currently only binomial and gaussian are supported).}
}
\value{
A scalar giving the model fit
}
\description{
Model fit function for the \code{ml_diag} function.
}
\details{
Returns either the square correlation between the dependent variable and its
predictions or the expected proportional reduction in error if the model is
a binomial GLM.
}
