% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_diag.R
\name{\%>\%}
\alias{\%>\%}
\alias{bm_shrink}
\title{bartMachine shrinkage function}
\usage{
bm_shrink(obj, train, test, ...)
}
\arguments{
\item{obj}{An object of class \code{lm} or \code{glm} with \code{family=binomial}.}

\item{train}{A vector of observations to be used in the training set.}

\item{test}{A vector of observations to be used in the testing set.}

\item{...}{Other arguments to be passed down to the \code{bartMachine} function.}
}
\value{
A list which contains the predicted values for the training set, predicted values for the
testing set and the \code{bartMachine} model used to generate the predictions.
}
\description{
bartMachine shrinkage function for the \code{ml_diag} function.
}
