% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_diag.R
\name{rf_shrink}
\alias{rf_shrink}
\title{Random forest shrinkage function}
\usage{
rf_shrink(obj, train, test, ...)
}
\arguments{
\item{obj}{An object of class \code{lm} or \code{glm} with \code{family=binomial}.}

\item{train}{A vector of observations to be used in the training set.}

\item{test}{A vector of observations to be used in the testing set.}

\item{...}{Other arguments to be passed down to the \code{randomForest} function.}
}
\value{
A list which contains the predicted values for the training set, predicted values for the
testing set and the \code{randomForest} model used to generate the predictions.
}
\description{
Random forest shrinkage function for the \code{ml_diag} function.
}
